Fix cast warning during compilation on amd64
git-svn-id: https://svn.code.sf.net/p/squirrelmail/code/trunk@14567 7612ce4b-ef26-0410-bec9-ea0150e637f0

+++ imap_proxy/src/main.c
@@ -283,7 +283,7 @@ int main( int argc, char *argv[] )
     const char *fn = "main()";
     char f_randfile[ PATH_MAX ];
     int listensd;                      /* socket descriptor we'll bind to */
-    int clientsd;                      /* incoming socket descriptor */
+    long clientsd;                     /* incoming socket descriptor */
     int sockaddrlen;                       
     struct sockaddr_storage srvaddr;
     struct sockaddr_storage cliaddr;
